Split EventBus and UniqueId + tests
